Implement flash messages
Flash messages are only implemented on profiles page for successful
submission of create profile. Other flashed messages implemented using
the passing of error messages to render_template function.

There was an issue with flash messages when reloading create profile
endpoint(/profile), which did not show flash messages on the pages were
reload. A createTemplate private function was implemented to get around
this issue, in which an error type and error message is passed to render
the create profile endpoint.

Create flash_messages.html template to keep flash message logic in one
place, as well as, included this template into the profiles.html
template.